+++ src/main/java/com/example/autopartz/model/Order.java
@@ -1,9 +1,12 @@
 package com.example.autopartz.model;
 
-import jakarta.persistence.*;
-import lombok.*;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
 import org.hibernate.Hibernate;
 
+import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
@@ -16,7 +19,7 @@ import java.util.Objects;
 @Table(name = "order_table")
 public class Order {
     @Id
-    Long ID_order;
+    Integer ID_order;
     String order_status;
     LocalDateTime order_date;
     @JoinColumn(name = "id_user")
