+++ src/main/java/com/example/autopartz/model/PartManufacturer.java
@@ -1,12 +1,15 @@
 package com.example.autopartz.model;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.*;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
 import org.hibernate.Hibernate;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import java.util.Objects;
 
 @Getter
@@ -17,7 +20,7 @@ import java.util.Objects;
 @Table(name = "part_manufacturer")
 public class PartManufacturer {
     @Id
-    Long ID_part_manufacturer;
+    Integer ID_part_manufacturer;
     @Column(name = "pm_name")
     String name;
     String pm_location;
