+++ src/main/java/com/example/autopartz/model/views/PartsForCarTypeAndCategory.java
@@ -1,11 +1,14 @@
 package com.example.autopartz.model.views;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.*;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
 import org.hibernate.annotations.Immutable;
 
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import java.io.Serial;
 import java.io.Serializable;
 
@@ -20,7 +23,7 @@ public class PartsForCarTypeAndCategory implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
     @Id
-    Long partid;
+    Integer partid;
     String partname;
     String cartype;
     String category;
