+++ src/main/java/com/example/autopartz/repository/OrdersForUserRepository.java
@@ -2,14 +2,14 @@ package com.example.autopartz.repository;
 
 import com.example.autopartz.model.views.OrdersForUser;
 import com.example.autopartz.model.views.OrdersForUserId;
-import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Repository
 @Transactional
 public interface OrdersForUserRepository extends JpaRepository<OrdersForUser, OrdersForUserId> {
-    List<OrdersForUser> findAllByUserid(Long userid);
+    List<OrdersForUser> findAllByUserid(Integer userid);
 }
