+++ src/main/java/com/example/autopartz/repository/PartRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PartRepository extends JpaRepository<Part,Long> {
+public interface PartRepository extends JpaRepository<Part,Integer> {
 }
