+++ src/main/java/com/example/autopartz/repository/PartsForCarTypeAndCategoryRepository.java
@@ -1,14 +1,14 @@
 package com.example.autopartz.repository;
 
 import com.example.autopartz.model.views.PartsForCarTypeAndCategory;
-import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Repository
 @Transactional
-public interface PartsForCarTypeAndCategoryRepository extends JpaRepository<PartsForCarTypeAndCategory,Long> {
+public interface PartsForCarTypeAndCategoryRepository extends JpaRepository<PartsForCarTypeAndCategory,Integer> {
     List<PartsForCarTypeAndCategory> findAllByCartypeAndCategory(String cartype, String category);
 }
