+++ src/main/java/com/example/autopartz/repository/RepairShopReviewSummaryRepository.java
@@ -1,13 +1,14 @@
 package com.example.autopartz.repository;
 
 import com.example.autopartz.model.views.RepairShopReviewsSummary;
-import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import javax.transaction.Transactional;
 
 // извештај за сите сервиси и за сите број на reviews и avg рејтинг
 @Repository
 @Transactional
-public interface RepairShopReviewSummaryRepository extends JpaRepository<RepairShopReviewsSummary,Long> {
-    // List<RepairShopReviewsSummary> findAllByRsid(Long rsid);
+public interface RepairShopReviewSummaryRepository extends JpaRepository<RepairShopReviewsSummary,Integer> {
+    // List<RepairShopReviewsSummary> findAllByRsid(Integer rsid);
 }
